+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * `state_space_ed()` and `state_space_ch()` allow now also irregular time series (before an error message was returned).
 
+`state_space_ed()` includes now an argument `na_rm = TRUE` (set as default) to deal with missing values, i.e. cases with NAs are dropped from the dataset prior to the analysis but the returned tibble includes these years (showing NA in the `$ed` column).
+
 
 # INDperform 0.2.2